Add configuration for the db-url
Allow to configure the database "base" url, this is expected to be a url
where the root of the couchdb is accessible. The url can be mapped to
another machine via a reverse-proxy or some other method, but its vital
that below the url the couchdb-databases are accessible.

Also note: Currently no cross-domain-requests are supported by the client,
hence the url must be on the same domain/origin as the html/js for the app.